+++ AddonLocale/lang_default.lua
@@ -6,7 +6,7 @@ local L = INL_Addon.strings;
 L["Chat"] = {
 	["Loaded"] = "Loaded |cffff0000%s|r locale.",
 	["ReloadUI"] = "Reload your UI to load the |cffff0000%s|r locale.",
-	["WowheadLink"] = "Wowhead link:"
+	["WowheadLink"] = "Wowhead link:",
 }
 L["Lang"] = {
 	["deDE"] = "German",
@@ -18,15 +18,16 @@ L["Lang"] = {
 	["koKR"] = "Korean",
 	["ptBR"] = "Portuguese",
 	["ruRU"] = "Russian",
-	["zhTW"] = "Chinese"
+	["zhCN"] = "Chinese (simplified)",
+	["zhTW"] = "Chinese (traditional)",
 }
 L["Options"] = {
 	["ReloadUIButton"] = "Reload UI",
 	["ResetButton"] = "Reset default config",
 	["selectedLocale"] = "Language (An UI reload is required)",
 	["showTooltipLine"] = "Show the localized string on a line on the tooltip.",
-	["showTooltipTitle"] = "Show the localized string in the tooltip title."
+	["showTooltipTitle"] = "Show the localized string in the tooltip title.",
 }
 L["Tooltip"] = {
-	["MissingLocale"] = "Missing locale"
+	["MissingLocale"] = "Missing locale",
 }