+++ DEPLOY.sh
@@ -1,11 +1,11 @@
 #!/bin/bash
 ADDON_MAIN_FOLDER="ItemNameLocalized"
-ADDON_VERSION="v1.3"
-GAME_VERSION="7.1.5"
+ADDON_VERSION="v1.4"
+GAME_VERSION="7.2.5"
 
 FOLDERS=("AddonLocale" "ItemLocales")
 FILES=("Core.lua" "Preload.lua" "ItemNameLocalized.toc" "CHANGELOG.md" "README.md" "LICENSE")
-FINAL_FILE="${ADDON_MAIN_FOLDER}_${ADDON_VERSION}_${GAME_VERSION}.zip"
+FINAL_FILE="${ADDON_MAIN_FOLDER}_${ADDON_VERSION}_wow${GAME_VERSION}.zip"
 
 echo "--------------------------"
 echo "> Deploying "$FINAL_FILE
@@ -26,9 +26,9 @@ done
 
 echo "--------------------------"
 echo "> Deleting old zip"
-rm -f ./FINAL_FILE
+rm -f ./$FINAL_FILE
 echo "> Creating zip"
-zip -r ./FINAL_FILE ./$ADDON_MAIN_FOLDER/*
+zip -r ./$FINAL_FILE ./$ADDON_MAIN_FOLDER/*
 echo "> zip created"
 echo "--------------------------"
 read -p "> Deploy finished. Press [Enter]..."