+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### v1.4
 * Added ptPR (thanks to Canettieri!) and ruRU (thanks to DogmatX!) localizations.
+* Updated some addon localizations.
 * Added 7.2.5 items and updated some old ones.
 
 ### v1.3